remove backend and simplify frontend
