--- .dockerignore
@@ -1,4 +1,0 @@
-node_modules
-.auth_info_baileys
-uploads
-numbers.json
