--- Dockerfile
@@ -1,7 +1,0 @@
-FROM node:20
-WORKDIR /app
-COPY package*.json ./
-RUN npm ci --omit=dev
-COPY . .
-EXPOSE 3000
-CMD ["npm","start"]
