--- Procfile
@@ -1,1 +1,0 @@
-web: sh -c 'node index.js & python server.py'
