--- package.json
@@ -1,17 +1,0 @@
-{
-  "name": "cedbrasil-bot",
-  "version": "1.0.0",
-  "main": "index.js",
-  "scripts": {
-    "start": "node index.js"
-  },
-  "dependencies": {
-    "@hapi/boom": "^10.0.1",
-    "@whiskeysockets/baileys": "^6.7.18",
-    "csv-parse": "^5.6.0",
-    "express": "^4.18.2",
-    "multer": "^2.0.1",
-    "qrcode": "^1.5.4",
-    "xlsx": "^0.18.5"
-  }
-}
