--- public/login.html
@@ -1,26 +1,0 @@
-<!DOCTYPE html>
-<html lang="pt-BR">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Login</title>
-    <link rel="stylesheet" href="style.css">
-    <style>
-        body { display:flex; justify-content:center; align-items:center; height:100vh; background:#f5f5f5; font-family:Arial, sans-serif; }
-        .login-box { background:white; padding:2rem; border-radius:8px; box-shadow:0 4px 20px rgba(0,0,0,0.1); }
-        .login-box h2 { margin-bottom:1rem; }
-        .login-box input { width:100%; padding:0.5rem; margin-bottom:1rem; }
-        .login-box button { width:100%; padding:0.5rem; cursor:pointer; }
-    </style>
-</head>
-<body>
-    <div class="login-box">
-        <h2>Login</h2>
-        <form method="post" action="/login">
-            <input type="text" name="username" placeholder="Usuário" required>
-            <input type="password" name="password" placeholder="Senha" required>
-            <button type="submit">Entrar</button>
-        </form>
-    </div>
-</body>
-</html>
