+++ public/painel.js
@@ -1,91 +1,69 @@
-const tabDisparos = document.getElementById('tab-disparos');
-const tabLeads = document.getElementById('tab-leads');
-const disparosSection = document.getElementById('disparos');
-const leadsSection = document.getElementById('leads');
-const gruposDiv = document.getElementById('grupos');
-const participantesDiv = document.getElementById('participantes');
-const listaNumeros = document.getElementById('lista-numeros');
-const grupoNome = document.getElementById('grupo-nome');
+const SEND_BASE = 'https://whatsapptest-stij.onrender.com';
 
-function switchTab(tab) {
-  if (tab === 'disparos') {
-    tabDisparos.classList.add('active');
-    tabLeads.classList.remove('active');
-    disparosSection.classList.remove('hidden');
-    leadsSection.classList.add('hidden');
-  } else {
-    tabLeads.classList.add('active');
-    tabDisparos.classList.remove('active');
-    leadsSection.classList.remove('hidden');
-    disparosSection.classList.add('hidden');
-    loadGrupos();
+const disparoState = {
+  running: false,
+  paused: false,
+  index: 0,
+  numbers: [],
+  mensagem: '',
+  timer: null
+};
+
+function formatNumero(n) {
+  n = String(n).replace(/\D/g, '');
+  if (n.startsWith('55') && n.length === 13 && n[4] === '9') {
+    n = n.slice(0, 4) + n.slice(5);
+  }
+  return n;
+}
+
+async function sendToApi(numero, mensagem) {
+  const url = `${SEND_BASE}/send?para=${numero}&mensagem=${encodeURIComponent(mensagem)}`;
+  try {
+    await fetch(url);
+  } catch (err) {
+    console.error('Falha ao enviar', numero, err);
   }
 }
 
-tabDisparos.addEventListener('click', () => switchTab('disparos'));
-tabLeads.addEventListener('click', () => switchTab('leads'));
+function sendNext() {
+  if (!disparoState.running || disparoState.paused) return;
+  if (disparoState.index >= disparoState.numbers.length) {
+    disparoState.running = false;
+    return;
+  }
+  const num = disparoState.numbers[disparoState.index];
+  sendToApi(num, disparoState.mensagem).finally(() => {
+    disparoState.index++;
+    const delay = Math.floor(Math.random() * (210 - 30 + 1) + 30) * 1000;
+    disparoState.timer = setTimeout(sendNext, delay);
+  });
+}
 
-async function disparar(acao) {
+function iniciar() {
   const numeros = document.getElementById('numeros').value;
   const mensagem = document.getElementById('mensagem').value;
-  if (acao === 'iniciar') {
-    await fetch('/disparar', {
-      method: 'POST',
-      headers: {'Content-Type': 'application/json'},
-      body: JSON.stringify({ numeros, mensagem })
-    });
-  } else {
-    await fetch('/' + acao, { method: 'POST' });
-  }
+  disparoState.numbers = numeros.split(',').map(n => formatNumero(n.trim())).filter(Boolean);
+  disparoState.mensagem = mensagem;
+  disparoState.index = 0;
+  disparoState.paused = false;
+  disparoState.running = true;
+  clearTimeout(disparoState.timer);
+  sendNext();
 }
 
-document.getElementById('iniciar').addEventListener('click', () => disparar('disparar'));
-document.getElementById('pausar').addEventListener('click', () => disparar('pausar'));
-document.getElementById('continuar').addEventListener('click', () => disparar('continuar'));
-
-async function loadGrupos() {
-  gruposDiv.innerHTML = 'Carregando...';
-  const res = await fetch('/grupos');
-  const grupos = await res.json();
-  gruposDiv.innerHTML = '';
-  grupos.forEach(g => {
-    const btn = document.createElement('button');
-    btn.textContent = g.nome;
-    btn.addEventListener('click', () => loadGrupo(g.nome));
-    gruposDiv.appendChild(btn);
-  });
+function pausar() {
+  if (!disparoState.running) return;
+  disparoState.paused = true;
+  clearTimeout(disparoState.timer);
 }
 
-async function loadGrupo(nome) {
-  const res = await fetch('/grupos/' + encodeURIComponent(nome));
-  const data = await res.json();
-  participantesDiv.classList.remove('hidden');
-  grupoNome.textContent = data.nome;
-  listaNumeros.innerHTML = '';
-  data.participantes.forEach(p => {
-    const li = document.createElement('li');
-    li.textContent = p.numero + (p.admin ? ' (admin)' : '');
-    li.dataset.numero = p.numero;
-    listaNumeros.appendChild(li);
-  });
+function continuar() {
+  if (!disparoState.running || !disparoState.paused) return;
+  disparoState.paused = false;
+  sendNext();
 }
 
-document.getElementById('exportar-csv').addEventListener('click', () => {
-  const rows = Array.from(listaNumeros.children)
-    .map(li => [li.dataset.numero || li.textContent]);
-  let csv = rows.map(r => r.join(',')).join('\n');
-  const blob = new Blob([csv], { type: 'text/csv' });
-  const a = document.createElement('a');
-  a.href = URL.createObjectURL(blob);
-  a.download = grupoNome.textContent + '.csv';
-  a.click();
-});
-
-document.getElementById('exportar-xlsx').addEventListener('click', () => {
-  const rows = Array.from(listaNumeros.children)
-    .map(li => [li.dataset.numero || li.textContent]);
-  const wb = XLSX.utils.book_new();
-  const ws = XLSX.utils.aoa_to_sheet(rows);
-  XLSX.utils.book_append_sheet(wb, ws, 'Leads');
-  XLSX.writeFile(wb, grupoNome.textContent + '.xlsx');
-});
+document.getElementById('iniciar').addEventListener('click', iniciar);
+document.getElementById('pausar').addEventListener('click', pausar);
+document.getElementById('continuar').addEventListener('click', continuar);
